refactor(todos): extract todoStatusSchema to shared module
list-todos and update-todo imported the VTODO status enum from
create-todo, making sibling tools depend on each other. Move it to
todo-status.ts so the dependency runs shared->tools.
